// src/lib/api/schemas/ContentReportStatus.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * Elite API
 * A backend API for https://eliteskyblock.com/ that provides Hypixel Skyblock data.
<br><br>
Use of this API requires following the [Elite API TOS](https://eliteskyblock.com/apiterms). This API is not affiliated with Hypixel or Mojang.
 * OpenAPI spec version: admin-v1
 */

export type ContentReportStatus = (typeof ContentReportStatus)[keyof typeof ContentReportStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ContentReportStatus = {
	open: 'open',
	reviewed: 'reviewed',
	dismissed: 'dismissed',
} as const;