// src/lib/api/schemas/ContentReportTargetType.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * Elite API
 * A backend API for https://eliteskyblock.com/ that provides Hypixel Skyblock data.
<br><br>
Use of this API requires following the [Elite API TOS](https://eliteskyblock.com/apiterms). This API is not affiliated with Hypixel or Mojang.
 * OpenAPI spec version: admin-v1
 */

export type ContentReportTargetType = (typeof ContentReportTargetType)[keyof typeof ContentReportTargetType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ContentReportTargetType = {
	guide: 'guide',
	comment: 'comment',
} as const;